# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/services": {
      "post": {
        "operationId": "ServiceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Servicio creado exitosamente"
          },
          "409": {
            "description": "Ya existe un servicio con ese nombre"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Crear un nuevo servicio (Solo Prestador de Servicio)",
        "tags": [
          "Servicios"
        ]
      },
      "get": {
        "operationId": "ServiceController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de servicios obtenida exitosamente"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener todos los servicios",
        "tags": [
          "Servicios"
        ]
      }
    },
    "/services/active": {
      "get": {
        "operationId": "ServiceController_findActive",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de servicios activos obtenida exitosamente"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener servicios activos",
        "tags": [
          "Servicios"
        ]
      }
    },
    "/services/{id}": {
      "get": {
        "operationId": "ServiceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Servicio obtenido exitosamente"
          },
          "404": {
            "description": "Servicio no encontrado"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener un servicio por ID",
        "tags": [
          "Servicios"
        ]
      },
      "patch": {
        "operationId": "ServiceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Servicio actualizado exitosamente"
          },
          "404": {
            "description": "Servicio no encontrado"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Actualizar un servicio",
        "tags": [
          "Servicios"
        ]
      },
      "delete": {
        "operationId": "ServiceController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Servicio eliminado exitosamente"
          },
          "404": {
            "description": "Servicio no encontrado"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Eliminar un servicio",
        "tags": [
          "Servicios"
        ]
      }
    },
...
    "/reviews": {
      "post": {
        "operationId": "ReviewController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reseña creada exitosamente"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Crear una nueva reseña",
        "tags": [
          "Reseñas"
        ]
      },
      "get": {
        "operationId": "ReviewController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de reseñas obtenida exitosamente"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener todas las reseñas",
        "tags": [
          "Reseñas"
        ]
      }
    },
    "/reviews/by-service/{serviceId}": {
      "get": {
        "operationId": "ReviewController_findByService",
        "parameters": [
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reseñas del servicio obtenidas exitosamente"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener reseñas de un servicio",
        "tags": [
          "Reseñas"
        ]
      }
    },
    "/reviews/by-client/{clientId}": {
      "get": {
        "operationId": "ReviewController_findByClient",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reseñas del cliente obtenidas exitosamente"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener reseñas de un cliente (Solo Admin)",
        "tags": [
          "Reseñas"
        ]
      }
    },
    "/reviews/my-reviews": {
      "get": {
        "operationId": "ReviewController_findMyReviews",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Mis reseñas obtenidas exitosamente"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener mis reseñas",
        "tags": [
          "Reseñas"
        ]
      }
    },
    "/reviews/service-stats/{serviceId}": {
      "get": {
        "operationId": "ReviewController_getServiceStats",
        "parameters": [
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estadísticas obtenidas exitosamente"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener estadísticas de reseñas de un servicio",
        "tags": [
          "Reseñas"
        ]
      }
    },
    "/reviews/{id}": {
      "get": {
        "operationId": "ReviewController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reseña obtenida exitosamente"
          },
          "404": {
            "description": "Reseña no encontrada"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener una reseña por ID",
        "tags": [
          "Reseñas"
        ]
      },
      "patch": {
        "operationId": "ReviewController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reseña actualizada exitosamente"
          },
          "404": {
            "description": "Reseña no encontrada"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Actualizar una reseña",
        "tags": [
          "Reseñas"
        ]
      },
      "delete": {
        "operationId": "ReviewController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reseña eliminada exitosamente"
          },
          "404": {
            "description": "Reseña no encontrada"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Eliminar una reseña (Solo Admin)",
        "tags": [
          "Reseñas"
        ]
      }
    },
...
  },
  "info": {
    "title": "Salon Backend API",
    "description": "API completa para el sistema de salon con autenticacion JWT",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del servicio",
            "example": "Corte de cabello"
          },
          "description": {
            "type": "string",
            "description": "Descripción del servicio",
            "example": "Corte de cabello profesional con lavado incluido"
          },
          "price": {
            "type": "number",
            "description": "Precio del servicio",
            "example": 25
          },
          "durationMinutes": {
            "type": "number",
            "description": "Duración del servicio en minutos",
            "example": 45
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado activo del servicio",
            "example": true,
            "default": true
          },
          "providerId": {
            "type": "string",
            "description": "ID del prestador asignado (solo administradores)",
            "example": "uuid-del-prestador"
          }
        },
        "required": [
          "name",
          "description",
          "price",
          "durationMinutes"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del servicio",
            "example": "Corte de cabello"
          },
          "description": {
            "type": "string",
            "description": "Descripción del servicio",
            "example": "Corte de cabello profesional con lavado incluido"
          },
          "price": {
            "type": "number",
            "description": "Precio del servicio",
            "example": 25
          },
          "durationMinutes": {
            "type": "number",
            "description": "Duración del servicio en minutos",
            "example": 45
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado activo del servicio",
            "example": true,
            "default": true
          },
          "providerId": {
            "type": "string",
            "description": "ID del prestador asignado (solo administradores)",
            "example": "uuid-del-prestador"
          }
        }
      },
...
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "serviceId": {
            "type": "string",
            "description": "ID del servicio",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "appointmentId": {
            "type": "string",
            "description": "ID de la cita (opcional)",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "rating": {
            "type": "number",
            "description": "Calificación (1-5)",
            "example": 5,
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "description": "Comentario de la reseña",
            "example": "Excelente servicio, muy profesional"
          }
        },
        "required": [
          "serviceId",
          "rating"
        ]
      },
      "UpdateReviewDto": {
        "type": "object",
        "properties": {
          "serviceId": {
            "type": "string",
            "description": "ID del servicio",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "appointmentId": {
            "type": "string",
            "description": "ID de la cita (opcional)",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "rating": {
            "type": "number",
            "description": "Calificación (1-5)",
            "example": 5,
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "description": "Comentario de la reseña",
            "example": "Excelente servicio, muy profesional"
          }
        }
      },
...
    }
  }
}